+++ test/Validators/HasOneDocumentValidatorTest.php
@@ -49,6 +49,19 @@ JSON_API;
         $validator = $this->hasOne();
 
         $this->assertTrue($validator->isValid($document));
+
+        return $document;
+    }
+
+    /**
+     * @param Document $document
+     * @depends testValid
+     */
+    public function testValidWithDefaultValidator(Document $document)
+    {
+        $validator = $this->factory->relationshipDocument();
+
+        $this->willExist()->assertTrue($validator->isValid($document));
     }
 
     public function testValidPolymorph()
@@ -337,7 +350,7 @@ JSON_API;
         $acceptable = null,
         $expectedType = 'users'
     ) {
-        $this->store->method('exists')->willReturn($exists);
+        $this->willExist($exists);
         $validator = $this->factory->hasOne($expectedType, $allowEmpty, $acceptable);
 
         return $this->factory->relationshipDocument($validator);